[Refactor] ui_interface.h to guiinterface.h
