+++ src/noui.cpp
@@ -7,7 +7,7 @@
 
 #include "noui.h"
 
-#include "ui_interface.h"
+#include "guiinterface.h"
 #include "util.h"
 
 #include <cstdio>
