+++ src/obfuscation.cpp
@@ -10,7 +10,7 @@
 #include "masternodeman.h"
 #include "script/sign.h"
 #include "swifttx.h"
-#include "ui_interface.h"
+#include "guiinterface.h"
 #include "util.h"
 #include <boost/algorithm/string/replace.hpp>
 #include <boost/filesystem.hpp>
