+++ src/qt/sendcoinsdialog.cpp
@@ -19,7 +19,7 @@
 
 #include "base58.h"
 #include "coincontrol.h"
-#include "ui_interface.h"
+#include "guiinterface.h"
 #include "utilmoneystr.h"
 #include "wallet/wallet.h"
