+++ src/zdogecchain.cpp
@@ -6,7 +6,7 @@
 #include "invalid.h"
 #include "main.h"
 #include "txdb.h"
-#include "ui_interface.h"
+#include "guiinterface.h"
 
 // 6 comes from OPCODE (1) + vch.size() (1) + BIGNUM size (4)
 #define SCRIPT_OFFSET 6
